profileScreen 005: update - * Updated User Details screen to only accept user id * Updated tests to work with the new parameter requirement * Founded new tests for invalid login attempts * Founded new tests for invalid registration attempts * Updated DAO to allow for passing only uid for users to be found * Implemented user verification on Login screen * Updated creating user within User Details view model

+++ app/src/main/java/com/dbad/justintime/f_login_register/presentation/register/RegisterScreen.kt
@@ -34,7 +34,7 @@ import com.dbad.justintime.ui.theme.JustInTimeTheme
 fun RegisterScreen(
     viewModel: RegisterViewModel,
     onCancelRegistration: () -> Unit,
-    onRegistration: () -> Unit,
+    onRegistration: (Int) -> Unit,
     modifier: Modifier = Modifier
 ) {
     val state by viewModel.state.collectAsState()
